server/requirements.txt
pymongo
mongoengine
fastapi
fastapi-cors
uvicorn
python-dotenv
pydantic
passlib[bcrypt]
firebase-admin
python-jose[cryptography]
python-multipart
email-validator